+++ tests/MediaQuery/SemanticLogMediaQueryLoggerTest.php
@@ -26,7 +26,6 @@ use function restore_error_handler;
 use function set_error_handler;
 
 use const E_USER_WARNING;
-use const JSON_PRESERVE_ZERO_FRACTION;
 use const JSON_THROW_ON_ERROR;
 
 /**
@@ -155,7 +154,7 @@ final class SemanticLogMediaQueryLoggerTest extends TestCase
     {
         /** @var array{open: list<array<array-key, mixed>>} */
         return json_decode(
-            json_encode($log, JSON_THROW_ON_ERROR | JSON_PRESERVE_ZERO_FRACTION),
+            json_encode($log, JSON_THROW_ON_ERROR),
             true,
             512,
             JSON_THROW_ON_ERROR,